+++ training_web_react/src/index.js
@@ -4,6 +4,7 @@ import './index.css';
 import App from './Containers/MainPage/App';
 import reportWebVitals from './reportWebVitals';
 
+
 ReactDOM.render(
     <App />,
   document.getElementById('root')
